+++ src/elements/Redirect.php
@@ -16,8 +16,11 @@ use craft\elements\actions\Restore;
 use craft\elements\db\ElementQueryInterface;
 use craft\helpers\Html;
 use craft\helpers\UrlHelper;
+use craft\models\Site;
 use craft\validators\DateTimeValidator;
 use craft\validators\SiteIdValidator;
+use craft\validators\UriValidator;
+use craft\validators\UrlValidator;
 use craft\web\ErrorHandler;
 use Throwable;
 use Twig\Error\LoaderError;
@@ -84,9 +87,9 @@ class Redirect extends Element
     public $destinationElementId;
 
     /**
-     * @var int|null destinationElementSiteId
+     * @var int|null destinationSiteId
      */
-    public $destinationElementSiteId;
+    public $destinationSiteId;
 
     /**
      * @inheritdoc
@@ -278,23 +281,31 @@ class Redirect extends Element
     }
 
     /**
-     * @return string|null
+     * Gets the actual final absolute destination URL
+     *
+     * @return string
+     * @throws \yii\base\Exception
      */
     public function getDestinationUrl()
     {
+        // Redirect to a site element
         if ($this->destinationElementId) {
-            $element = Craft::$app->elements->getElementById($this->destinationElementId, null, $this->destinationElementSiteId ?? $this->siteId);
+            $element = Craft::$app->elements->getElementById($this->destinationElementId, null, $this->destinationSiteId ?? $this->siteId);
             if ($element && $element->getUrl()) {
                 return $element->getUrl();
             }
-        } elseif ($this->destinationUrl) {
+            // We don't have an element, but we do have a site ID, so send to that site, regardless of URL format
+        } elseif (isset($this->destinationUrl, $this->destinationSiteId)) {
+            // UrlHelper::siteUrl() will try to default to absolute URLs, so we'll handle it ourselves
+            return $this->getDestinationSite()->getBaseUrl() . $this->destinationUrl;
+        } else {
+            // No site ID, so if it's absolute, send to that URL
             if (UrlHelper::isAbsoluteUrl($this->destinationUrl)) {
                 return $this->destinationUrl;
             }
-
-            return UrlHelper::siteUrl($this->destinationUrl, null, null, $this->destinationElementSiteId ?? $this->siteId);
+            // It's not absolute, so use the site the redirect was saved in
+            return $this->getSite()->getBaseUrl() . $this->destinationUrl;
         }
-        return null;
     }
 
     /**
@@ -316,6 +327,18 @@ class Redirect extends Element
     }
 
     /**
+     * Gets the destination site if one is set
+     *
+     * @return Site|null
+     */
+    public function getDestinationSite() {
+        if($this->destinationSiteId === null) {
+            return null;
+        }
+        return Craft::$app->sites->getSiteById($this->destinationSiteId);
+    }
+
+    /**
      * Use the sourceUrl as the string representation.
      *
      * @return string
@@ -327,18 +350,28 @@ class Redirect extends Element
     {
         $rules = parent::defineRules();
         $rules[] = [['hitAt'], DateTimeValidator::class];
-        $rules[] = [['hitCount', 'destinationElementId', 'destinationElementSiteId'], 'number', 'integerOnly' => true];
-        $rules[] = [['destinationElementSiteId'], SiteIdValidator::class];
+        $rules[] = [['hitCount', 'destinationElementId', 'destinationSiteId'], 'number', 'integerOnly' => true];
+        $rules[] = [['sourceUrl', 'destinationUrl'], 'string', 'max' => 255];
+        $rules[] = [['sourceUrl', 'type'], 'required'];
+
+        $rules[] = [['type'], 'in', 'range' => [self::TYPE_STATIC, self::TYPE_DYNAMIC]];
+        $rules[] = [['statusCode'], 'in', 'range' => array_keys(self::STATUS_CODE_OPTIONS)];
+
+        $rules[] = [['destinationSiteId'], SiteIdValidator::class];
+        $rules[] = ['destinationElementId', 'exist', 'targetClass' => \craft\records\Element::class];
+        $rules[] = ['destinationSiteId', 'required', 'when' => function($model) {
+            return !empty($model->destinationElementId);
+        }];
         $rules[] = ['destinationUrl', 'required', 'when' => function($model) {
             return empty($model->destinationElementId);
         }];
-        $rules[] = ['destinationElementSiteId', 'required', 'when' => function($model) {
-            return !empty($model->destinationElementId);
+        $rules[] = ['destinationUrl', UrlValidator::class, 'when' => function($model) {
+            return empty($model->destinationSiteId);
         }];
-        $rules[] = [['sourceUrl', 'destinationUrl'], 'string', 'max' => 255];
-        $rules[] = [['sourceUrl', 'type'], 'required'];
-        $rules[] = [['type'], 'in', 'range' => [self::TYPE_STATIC, self::TYPE_DYNAMIC]];
-        $rules[] = [['statusCode'], 'in', 'range' => array_keys(self::STATUS_CODE_OPTIONS)];
+        $rules[] = ['destinationUrl', UriValidator::class, 'when' => function($model) {
+            return !empty($model->destinationSiteId);
+        }];
+
         return $rules;
     }
 
@@ -368,16 +401,17 @@ class Redirect extends Element
      */
     public function afterSave(bool $isNew)
     {
-        // Get the redirect record
-        if (!$isNew) {
-            $record = RedirectRecord::findOne($this->id);
+        if (!$this->propagating) {
+            if (!$isNew) {
+                $record = RedirectRecord::findOne($this->id);
 
-            if (!$record) {
-                throw new Exception('Invalid redirect ID: ' . $this->id);
+                if (!$record) {
+                    throw new Exception('Invalid redirect ID: ' . $this->id);
+                }
+            } else {
+                $record = new RedirectRecord();
+                $record->id = (int)$this->id;
             }
-        } else {
-            $record = new RedirectRecord();
-            $record->id = $this->id;
 
             if ($this->hitCount > 0) {
                 $record->hitCount = $this->hitCount;
@@ -390,31 +424,31 @@ class Redirect extends Element
             } else {
                 $record->hitAt = null;
             }
-        }
 
-        $record->sourceUrl = $this->formatUrl(trim($this->sourceUrl), true);
-        if ($this->destinationUrl) {
-            $record->destinationUrl = $this->formatUrl(trim($this->destinationUrl), false);
-        }
+            $record->sourceUrl = $this->formatUrl(trim($this->sourceUrl), true);
+            if ($this->destinationUrl) {
+                $record->destinationUrl = $this->formatUrl(trim($this->destinationUrl), false);
+            }
 
-        if ($this->destinationElementId) {
-            $record->destinationElementId = $this->destinationElementId;
-        }
+            if ($this->destinationElementId) {
+                $record->destinationElementId = $this->destinationElementId;
+            }
 
-        if ($this->destinationElementSiteId) {
-            $record->destinationElementSiteId = $this->destinationElementSiteId;
-        }
+            if ($this->destinationSiteId) {
+                $record->destinationSiteId = $this->destinationSiteId;
+            }
 
-        $record->statusCode = $this->statusCode;
-        $record->type = $this->type;
-        if ($this->dateCreated) {
-            $record->dateCreated = $this->dateCreated;
-        }
-        if ($this->dateUpdated) {
-            $record->dateUpdated = $this->dateUpdated;
-        }
+            $record->statusCode = $this->statusCode;
+            $record->type = $this->type;
+            if ($this->dateCreated) {
+                $record->dateCreated = $this->dateCreated;
+            }
+            if ($this->dateUpdated) {
+                $record->dateUpdated = $this->dateUpdated;
+            }
 
-        $record->save(false);
+            $record->save(false);
+        }
         parent::afterSave($isNew);
     }
 
@@ -535,7 +569,7 @@ class Redirect extends Element
     {
         if (isset($this->destinationElementId)) {
             return Craft::$app->getView()->renderTemplate('_elements/element', [
-                'element' => Craft::$app->elements->getElementById($this->destinationElementId, null, $this->destinationElementSiteId),
+                'element' => Craft::$app->elements->getElementById($this->destinationElementId, null, $this->destinationSiteId),
             ]);
         }
         if ($this->destinationUrl) {